% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_scene.R
\name{coords_folder}
\alias{coords_folder}
\title{Get mean coordinates of all patches from a folder #'
(Description paragraph)}
\usage{
coords_folder(season, scene_id, dat = "lc")
}
\arguments{
\item{season}{The season of this scene.}

\item{scene_id}{The scene index.}

\item{dat}{The data type of options ('lc','s1','s2')}
}
\value{
A vector of (longitude, latitude).
}
\description{
Get mean coordinates of all patches from a folder #'
(Description paragraph)
}
\details{
(Details section)
}
\examples{
# (Examples section)
 #'
}
