% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_patch.R
\name{coords_patch}
\alias{coords_patch}
\title{Get coordinates from a geographical patch #'
(Description paragraph)}
\usage{
coords_patch(filename)
}
\arguments{
\item{filename}{The directory of the geographical patch.}
}
\value{
A vector of (longitude, latitude).
}
\description{
Get coordinates from a geographical patch #'
(Description paragraph)
}
\details{
(Details section)
}
\examples{
# (Examples section)
 #'
}
