% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_flow.R
\name{parse_path}
\alias{parse_path}
\title{Paths to Sentinel Data}
\usage{
parse_path(path)
}
\arguments{
\item{path}{The full path to a BigEarth sample.}
}
\value{
result A data.frame containing the original path and the sample's
attributes.
}
\description{
This is a helper to extract seasonality and region information from the path
to a sentinel image patch in BigEarthNet.
}
