% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_dirs.R
\docType{data}
\name{scene_dirs}
\alias{scene_dirs}
\title{Scene Directories of Sen12MS Dataset}
\format{
scene_dirs is a vector consisting of all the 756 scene directories.

}
}
\source{
The original dataset comes from
(\url{https://arxiv.org/abs/1906.07789})
}
\usage{
scene_dirs
}
\description{
Scene Directories of Sen12MS Dataset
}
\keyword{datasets}
