% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_dir.R
\name{season_dir}
\alias{season_dir}
\title{Get season directories#'
(Description paragraph)}
\usage{
season_dir(season, base_dir = "/Volumes/ksankaran/Sen12MS")
}
\arguments{
\item{season}{The season.}

\item{base_dir}{The base directory.}
}
\value{
A directory of files in this season.
}
\description{
Get season directories#'
(Description paragraph)
}
\details{
(Details section)
}
\examples{
# (Examples section)
 #'
}
