% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Subsampling the Sen12MS dataset #'
(Description paragraph)}
\usage{
subsample(n, n_per_scene, dirs)
}
\arguments{
\item{n}{The directory of the geographical patch.}
}
\value{
A dataframe of metadata.
}
\description{
Subsampling the Sen12MS dataset #'
(Description paragraph)
}
\details{
(Details section)
}
\examples{
# (Examples section)
 #'
}
