% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svcca.R
\name{svcca}
\alias{svcca}
\title{Simple SVCCA Implementation}
\usage{
svcca(h1, h2, K = 25, L = 8)
}
\arguments{
\item{h1}{The first matrix of activations.}

\item{h2}{The second matrix of activations}

\item{K}{The number of SVD directions to project down onto.}

\item{L}{The number of CCA rho\link{k}'s to evaluate.}
}
\description{
A SVCCA implementation that assumes all neuron activations have already been
subsampled / rearranged into matrices. Essentially a wrapper around the CCA
in the PMA package. Note that this will actually be a regularized CCA, which
is different from the original definition.
}
